Many fixes
- Fix style loading
- Revert to using SearchEntry due to focus issues
- Move menu into window.ui
- Move actions around a bit
- Let gtk/libadwaita handle the shortcuts window
- Use action names instead of key combos in the shortcuts window
