+++ build-aux/flatpak/com.github.fushinari.Wordbook.Devel.json
@@ -8,6 +8,7 @@
     "--share=network",
     "--socket=pulseaudio",
     "--share=ipc",
+    "--device=dri",
     "--socket=fallback-x11",
     "--socket=wayland"
   ],
