--- data/resources/ui/menu.ui
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<interface>
-  <requires lib="gtk" version="4.0"/>
-  <menu id="wordbook-menu">
-    <section>
-      <attribute name="id">search-section</attribute>
-      <item>
-        <attribute name="label" translatable="yes">Paste &amp; Search</attribute>
-        <attribute name="action">app.paste-search</attribute>
-      </item>
-      <item>
-        <attribute name="label" translatable="yes">Search Selected Text</attribute>
-        <attribute name="action">app.search-selected</attribute>
-      </item>
-      <item>
-        <attribute name="label" translatable="yes">_Random Word</attribute>
-        <attribute name="action">app.random-word</attribute>
-      </item>
-    </section>
-    <section>
-      <attribute name="id">edit-section</attribute>
-      <item>
-        <attribute name="label" translatable="yes">_Preferences</attribute>
-        <attribute name="action">app.preferences</attribute>
-      </item>
-      <attribute name="id">help-section</attribute>
-      <item>
-        <attribute name="label" translatable="yes">_Keyboard Shortcuts</attribute>
-        <attribute name="action">app.shortcuts</attribute>
-      </item>
-      <item>
-        <attribute name="label" translatable="yes">_About Wordbook</attribute>
-        <attribute name="action">app.about</attribute>
-      </item>
-    </section>
-  </menu>
-</interface>
