+++ data/resources/ui/window.ui
@@ -2,6 +2,38 @@
 <interface>
   <requires lib="gtk" version="4.0"/>
   <requires lib="libadwaita" version="1.0"/>
+  <menu id="app-menu">
+    <section>
+      <attribute name="id">search-section</attribute>
+      <item>
+        <attribute name="label" translatable="yes">Paste &amp; Search</attribute>
+        <attribute name="action">win.paste-search</attribute>
+      </item>
+      <item>
+        <attribute name="label" translatable="yes">Search Selected Text</attribute>
+        <attribute name="action">win.search-selected</attribute>
+      </item>
+      <item>
+        <attribute name="label" translatable="yes">_Random Word</attribute>
+        <attribute name="action">win.random-word</attribute>
+      </item>
+    </section>
+    <section>
+      <attribute name="id">help-section</attribute>
+      <item>
+        <attribute name="label" translatable="yes">_Preferences</attribute>
+        <attribute name="action">win.preferences</attribute>
+      </item>
+      <item>
+        <attribute name="label" translatable="yes">_Keyboard Shortcuts</attribute>
+        <attribute name="action">win.show-help-overlay</attribute>
+      </item>
+      <item>
+        <attribute name="label" translatable="yes">_About Wordbook</attribute>
+        <attribute name="action">app.about</attribute>
+      </item>
+    </section>
+  </menu>
   <template class="WordbookWindow" parent="AdwApplicationWindow">
     <property name="default-width">390</property>
     <property name="default-height">610</property>
@@ -21,7 +53,7 @@
                     <property name="hexpand">True</property>
                     <property name="spacing">4</property>
                     <child>
-                      <object class="GtkEntry" id="search_entry">
+                      <object class="GtkSearchEntry" id="search_entry">
                         <property name="hexpand">True</property>
                         <property name="activates-default">True</property>
                       </object>
@@ -47,6 +79,7 @@
             </child>
             <child type="end">
               <object class="GtkMenuButton" id="wordbook_menu_button">
+                <property name="menu-model">app-menu</property>
                 <property name="receives-default">True</property>
                 <property name="direction">none</property>
               </object>
@@ -224,7 +257,7 @@
                                                 <property name="hexpand">False</property>
                                                 <child>
                                                   <object class="GtkLabel" id="term_view">
-                                                    <property name="label" translatable="no">&lt;span size=&quot;large&quot; weight=&quot;bold&quot;&gt;Term&lt;/span&gt;</property>
+                                                    <property name="label" translatable="no">Term&gt;</property>
                                                     <property name="use-markup">True</property>
                                                     <property name="single-line-mode">True</property>
                                                     <property name="xalign">0</property>
@@ -233,7 +266,7 @@
                                                 </child>
                                                 <child>
                                                   <object class="GtkLabel" id="pronunciation_view">
-                                                    <property name="label" translatable="yes">&lt;i&gt;/Pronunciation/&lt;/i&gt;</property>
+                                                    <property name="label" translatable="yes">/Pronunciation/</property>
                                                     <property name="use-markup">True</property>
                                                     <property name="selectable">True</property>
                                                     <property name="single-line-mode">True</property>
