+++ po/POTFILES
@@ -1,4 +1,3 @@
-data/resources/ui/menu.ui
 data/resources/ui/settings_window.ui
 data/resources/ui/shortcuts_window.ui
 data/resources/ui/window.ui
