+++ wordbook/main.py
@@ -9,7 +9,7 @@ from gettext import gettext as _
 gi.require_version("Gdk", "4.0")
 gi.require_version("Gtk", "4.0")
 gi.require_version("Adw", "1")
-from gi.repository import Gio, GLib, Adw  # noqa
+from gi.repository import Adw, Gio, GLib, Gtk  # noqa
 
 from wordbook import base, utils  # noqa
 from wordbook.window import WordbookWindow  # noqa
@@ -32,6 +32,8 @@ class Application(Adw.Application):
             application_id=app_id,
             flags=Gio.ApplicationFlags.HANDLES_COMMAND_LINE,
         )
+        GLib.set_application_name(_("Wordbook"))
+        GLib.set_prgname(self.app_id)
 
         self.app_id = app_id
         self.version = version
@@ -62,40 +64,21 @@ class Application(Adw.Application):
             None,
         )
 
+        Adw.StyleManager.get_default().set_color_scheme(
+            Adw.ColorScheme.FORCE_DARK
+            if Settings.get().gtk_dark_ui
+            else Adw.ColorScheme.PREFER_LIGHT
+        )
+
+        base.fold_gen()
+
+    def do_startup(self):
+        """Manage startup of the application."""
+        self.set_resource_base_path(utils.RES_PATH)
+        Adw.Application.do_startup(self)
+
     def do_activate(self):
         """Activate the application."""
-
-        def setup_actions(window):
-            """Setup the Gio actions for the application."""
-            about_action = Gio.SimpleAction.new("about", None)
-            about_action.connect("activate", window.on_about)
-            self.add_action(about_action)
-
-            paste_search_action = Gio.SimpleAction.new("paste-search", None)
-            paste_search_action.connect("activate", window.on_paste_search)
-            self.add_action(paste_search_action)
-
-            preferences_action = Gio.SimpleAction.new("preferences", None)
-            preferences_action.connect("activate", window.on_preferences)
-            self.add_action(preferences_action)
-
-            random_word_action = Gio.SimpleAction.new("random-word", None)
-            random_word_action.connect("activate", window.on_random_word)
-            self.add_action(random_word_action)
-
-            search_selected_action = Gio.SimpleAction.new("search-selected", None)
-            search_selected_action.connect("activate", window.on_search_selected)
-            self.add_action(search_selected_action)
-
-            shortcuts_action = Gio.SimpleAction.new("shortcuts", None)
-            shortcuts_action.connect("activate", window.on_shortcuts)
-            self.add_action(shortcuts_action)
-
-            self.set_accels_for_action("app.search-selected", ["<Primary>s"])
-            self.set_accels_for_action("app.random-word", ["<Primary>r"])
-            self.set_accels_for_action("app.paste-search", ["<Primary><Shift>v"])
-            self.set_accels_for_action("app.preferences", ["<Primary>comma"])
-
         self.win = self.get_active_window()
         if not self.win:
             self.win = WordbookWindow(
@@ -103,7 +86,7 @@ class Application(Adw.Application):
                 title=_("Wordbook"),
                 term=self.lookup_term,
             )
-            setup_actions(self.win)
+            self.setup_actions()
 
         self.win.present()
 
@@ -129,13 +112,29 @@ class Application(Adw.Application):
         self.activate()
         return 0
 
-    def do_startup(self):
-        """Manage startup of the application."""
-        Adw.Application.do_startup(self)
-        Adw.StyleManager.get_default().set_color_scheme(
-            Adw.ColorScheme.FORCE_DARK
-            if Settings.get().gtk_dark_ui
-            else Adw.ColorScheme.PREFER_LIGHT
+    def on_about(self, _action, _param):
+        """Show the about window."""
+        about_dialog = Gtk.AboutDialog(transient_for=self.win, modal=True)
+        about_dialog.set_logo_icon_name(Gio.Application.get_default().app_id)
+        about_dialog.set_program_name(_("Wordbook"))
+        about_dialog.set_version(Gio.Application.get_default().version)
+        about_dialog.set_comments(_("Lookup definitions of any English term."))
+        about_dialog.set_authors(
+            [
+                "Mufeed Ali",
+            ]
         )
+        about_dialog.set_license_type(Gtk.License.GPL_3_0)
+        about_dialog.set_website("https://www.github.com/fushinari/wordbook")
+        about_dialog.set_copyright(_("Copyright © 2016–2021 Mufeed Ali"))
+        about_dialog.present()
 
-        base.fold_gen()
+    def setup_actions(self):
+        about_action = Gio.SimpleAction.new("about", None)
+        about_action.connect("activate", self.on_about)
+        self.add_action(about_action)
+
+        self.set_accels_for_action("win.search-selected", ["<Primary>s"])
+        self.set_accels_for_action("win.random-word", ["<Primary>r"])
+        self.set_accels_for_action("win.paste-search", ["<Primary><Shift>v"])
+        self.set_accels_for_action("win.preferences", ["<Primary>comma"])
